requirements.txt
numpy==1.19.4
pandas==1.3.5
pyarrow==2.0.0

matplotlib==3.3.3
seaborn==0.12.1

scipy==1.5.4
jupyterlab==3.1.7

google-cloud-bigquery==2.4.0
google-cloud-storage==1.31.2
